--- Blog/backend/db.js
@@ -1,10 +1,0 @@
-import mongoose from "mongoose";
-
-export const connectionToDB  = async () => {
-    try {
-        await mongoose.connect("mongodb://localhost/technical-test");
-        console.log("DB is connected");
-    } catch (error) {
-        console.log(error);
-    }
-};
