--- Blog/backend/middlewares/validator.middleware.js
@@ -1,10 +1,0 @@
-export const validateSchema = (schema) => (req, res, next) => {
-    try {
-        schema.parse(req.body);
-        next();
-    } catch (error) {
-        return res
-            .status(400)
-            .json({ error: error.errors.map(error => error.message) })
-    }
-};
