--- Blog/backend/models/post.model.js
@@ -1,29 +1,0 @@
-import mongoose from "mongoose";
-
-const postSchema = new mongoose.Schema({
-    title: {
-        type: String,
-        required: true,
-    },
-    author: {
-        type: String,
-        required: true
-    },
-    date: {
-        type: String,
-        required: true
-    },
-    content: {
-        type: String,
-        required: true
-    },
-    user: {
-        type: mongoose.Schema.Types.ObjectId,
-        ref: "User",
-        required: true
-    }
-}, {
-    timestamps: true
-});
-
-export default mongoose.model("Post", postSchema);
