--- Blog/backend/models/user.model.js
@@ -1,25 +1,0 @@
-import mongoose from "mongoose";
-
-const userSchema = new mongoose.Schema({
-    username: {
-        type: String,
-        required: true,
-    },
-    lastName: {
-        type: String,
-        required: true,
-    },
-    email: {
-        type: String,
-        required: true,
-        unique: true,
-    },
-    password: {
-        type: String,
-        required: true,
-    },
-}, {
-    timestamps: true
-});
-
-export default mongoose.model("User", userSchema);
